[IMP] website_slides: Add elearning feature to slides
Purpose of this commit is to introduce a new way of displaying and managing
slides in eLearning module. Its purpose is to give a better experience to
users when going through a course. Notably a fullscreen mode allows to take
lessons one by one without going out of the elearning display. Integration
of all slide types eases taking the course step by step.

Quiz are introduced in this commit. Those are an addition to slides with
some question / answers (multiple choice). It allows customer to gain karma
and improves gamification. Certifications are still done using the survey
application. Quiz targets only small tests at the end of a given slide.

A new widget is added to display and control slide display in fullscreen
mode. Old display is still available for documentation channels or when
going out of fullscreen mode, to see details on comments / review / statistics
and have access to share options.

This commit is linked to task ID 1902304 and PR #29876. It closes the main
work on refactoring website slides into eLearning, with other tasks already
merged in current community [1][2][3][4][5].

[1] Task ID 1940360 landed at ac4f6abff27d97e27dbf1266b35aa5315f5ddd31: certification inclusion
[2] Task ID 1936153 landed at 24ff9485e8fcb9e29f0383c4fb5c7ff3a4b0ee50: homepage inclusion
[3] Task ID 1937160 landed at f7ea8c2a64625fb8d5b62c0d8c5bfb2ed6175725: selling courses
[4] Task ID 1922159 and 1940516 landed at b4e30dc9b43bc88753c24a987ea4f33d49431a1c and 3891e29f4a2e0662d96be971968a4f7787476357:
    new user profile and gamification of slides / forum
[5] Task ID 1938643 landed at 91f295231660bab205fd5a0b23f8b5beaafb8858: upload channel/slide