[FIX] hr_holidays: change selection keys from floats to integers
In the following commit (https://github.com/odoo/odoo/commit/e9a7e94ffa9d4da912ece1a3aea1cd9eef3b3aa9) we introduced some floats as keys
to Selection fields (namely request_hour_from and request_hour_to)

This lead to the ORM creating a column of type `int4` to store this
field which rounded the values.

In order to repair that faulty behaviour we use negative integer values
for the half hours, which is then transformed back into a float in the
code.

closes odoo/odoo#28877