[FIX] sale: de translation can ecommerce wired
There was an issue in translation that broke the mail template "Sales
Order: Send by email" in german. This prevented to do a wired sale on
the ecommerce in german.

This has been fixed in https://transifex.com/odoo/ but is done in this
PR to fix it more speedily than next week.

opw-1911419
opw-1911696
closes #29033