[FIX] hr_org_chart: 3 level num subordinate right
On the third level of a organization chart, the number of subordinates
on the right but over the container boundary.

For the second level there was already a special case:

  padding-left: $tmp-gap-base;

So for the third a fix could have been:

  padding-left: $tmp-gap-base*2;

But this changeset replaces margin by padding so invisible content does
not create a horizontal scrollbar (since we had 100% width with
margin-left that busted it out of its container).

opw-1912437
closes #29043