[IMP] hr_payroll: Remove hr.rule.input model
There are currently 2 models:
- hr.rule.input to link some inputs to an hr.payslip.rule
- hr.payslip.input to add some additional inputs on a payslip.

The first model is actually there to automatically add inputs on the
additional inputs on the payslip. This is problematic because these
rules are always added (For example a salary reduction due to a
police fine). As these inputs are always added on the payslip, the
the related rules are always computed with an amount of 0.

We decided to remove this mechanism for the following reasons:
- If the rule is important and common, then define the needed values on
  the contract, and compute the rules from that.
- If the rule is not so important, it should be added manually when needed.